Rebase against Netty4, all r2 tests pass

+++ r2/src/main/java/com/linkedin/r2/transport/http/client/HttpNettyClient.java
@@ -130,7 +130,7 @@ import org.slf4j.LoggerFactory;
                          long requestTimeout,
                          long idleTimeout,
                          long shutdownTimeout,
-                         int maxResponseSize,
+                         long maxResponseSize,
                          SSLContext sslContext,
                          SSLParameters sslParameters,
                          ExecutorService callbackExecutors,
